+++ app/src/main/java/com/example/goclass/ui/classui/chats/ChatViewModel.kt
@@ -1,20 +1,20 @@
-package com.example.goclass.ui.classui.chats.chat
+package com.example.goclass.ui.classui.chats
 
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.goclass.network.dataclass.AttendancesResponse
-import com.example.goclass.network.dataclass.Comments
 import com.example.goclass.network.dataclass.CommentsResponse
+import com.example.goclass.network.dataclass.Messages
+import com.example.goclass.network.dataclass.MessagesResponse
 import com.example.goclass.repository.ChatRepository
 import kotlinx.coroutines.launch
 
-class ChatCommentViewModel(
-    private val repository: ChatRepository
+class ChatViewModel(
+    private val repository: ChatRepository,
 ) : ViewModel() {
-    private val commentListLiveData : MutableLiveData<List<CommentsResponse>> = MutableLiveData()
+    private val messageListLiveData : MutableLiveData<List<MessagesResponse>> = MutableLiveData()
     private val _toastMessage = MutableLiveData<String>()
     private val _sendSuccess = MutableLiveData<Boolean>()
     private val _editSuccess = MutableLiveData<Boolean>()
@@ -23,81 +23,78 @@ class ChatCommentViewModel(
     private val sendSuccess: LiveData<Boolean> get() = _sendSuccess
     private val editSuccess: LiveData<Boolean> get() = _editSuccess
 
-    fun chatCommentSend(
+    fun chatChannelSend(
         classId: Int,
-        id: Int,
-        userId: Int,
+        senderId: Int,
         content: String,
     ) {
-        val sendComments = Comments(userId, content)
+        val sendMessages = Messages(senderId, content)
         viewModelScope.launch {
             try {
-                val response = repository.chatCommentSend(classId, id, sendComments)
-                if(response.code == 200){
-                    chatCommentGetList(classId, id)
+                val response = repository.chatChannelSend(classId, sendMessages)
+                if(response.code == 200) {
+                    chatChannelGetList(classId)
                     _toastMessage.postValue("Success")
                     _sendSuccess.postValue(true)
                 } else {
                     _toastMessage.postValue("Failure")
                     _sendSuccess.postValue(false)
                 }
-            } catch (e: Exception) {
-                Log.d("commentSendError", e.message.toString())
+            } catch (e:Exception) {
+                Log.d("chatSendError", e.message.toString())
                 _toastMessage.postValue("Error: ${e.message}")
                 _sendSuccess.postValue(false)
             }
         }
     }
 
-    fun chatCommentEdit(
+    fun chatChannelEdit(
         classId: Int,
         content: String,
-        commentId: Int,
         messageId: Int,
     ) {
-        val comments = Comments(content, messageId)
+        val editMessages = Messages(content, messageId)
         viewModelScope.launch {
             try {
-                val response = repository.chatCommentEdit(classId, commentId, comments)
-                if(response.code == 200){
-                    chatCommentGetList(classId, commentId)
+                val response = repository.chatChannelEdit(classId, editMessages)
+                if(response.code == 200) {
+                    chatChannelGetList(classId)
                     _toastMessage.postValue("Success")
                     _editSuccess.postValue(true)
                 } else {
                     _toastMessage.postValue("Failure")
                     _editSuccess.postValue(false)
                 }
-            } catch (e: Exception) {
-                Log.d("commentEditError", e.message.toString())
+            } catch (e:Exception) {
+                Log.d("chatEditError", e.message.toString())
                 _toastMessage.postValue("Error: ${e.message}")
                 _editSuccess.postValue(false)
             }
         }
     }
 
-    fun chatCommentGetList(
+    fun chatChannelGetList(
         classId: Int,
-        commentId: Int,
-    ): MutableLiveData<List<CommentsResponse>> {
+    ): MutableLiveData<List<MessagesResponse>> {
         viewModelScope.launch {
             try {
-                val response = repository.chatCommentGetList(classId, commentId)
+                val response = repository.chatChannelGetList(classId)
                 if(response.code == 200){
-                    commentListLiveData.postValue(response.commentList)
+                    messageListLiveData.postValue(response.messageList)
                     _toastMessage.postValue("Success")
                 } else {
                     _toastMessage.postValue("Failure")
                 }
             } catch (e:Exception) {
-                Log.d("commentListGetError", e.message.toString())
+                Log.d("chatListGetError", e.message.toString())
                 _toastMessage.postValue("Error: ${e.message}")
             }
         }
-        return commentListLiveData
+        return messageListLiveData
     }
 
-    fun accessCommentListLiveData(): MutableLiveData<List<CommentsResponse>> {
-        return commentListLiveData
+    fun accessMessageListLiveData(): MutableLiveData<List<MessagesResponse>> {
+        return messageListLiveData
     }
 
     fun accessToastMessage(): LiveData<String> {
