+++ app/src/main/res/drawable/appbar_bg.xml
@@ -1,20 +1,16 @@
-<?xml version="1.0" encoding="utf-8"?>
-<vector
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="360dp"
-    android:height="70dp"
+    android:height="56dp"
     android:viewportWidth="360"
-    android:viewportHeight="70"
-    >
-    <group>
-        <clip-path
-            android:pathData="M0 0H360V70H0V0Z"
-            />
-        <path
-            android:pathData="M0 0H360V70H0V0Z"
-            android:strokeWidth="2"
-            android:strokeColor="#D8D8D8"
-            />
-    </group>
+    android:viewportHeight="56">
+  <path
+      android:pathData="M0,0H360V56H0V0Z"
+      android:fillColor="#ffffff"/>
+  <group>
+    <clip-path
+        android:pathData="M0,0H360V56H0V0Z"/>
+    <path
+        android:pathData="M360,55H0V57H360V55Z"
+        android:fillColor="#F0F0F0"/>
+  </group>
 </vector>
