+++ app/src/main/res/drawable/chat_bg.xml
@@ -1,0 +1,16 @@
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape>
+            <solid android:color="#FFFFFF"/>
+        </shape>
+    </item>
+
+    <item android:top="-2dp" android:left="-2dp" android:right="-2dp">
+        <shape>
+            <solid android:color="#00000000"/>
+            <stroke
+                android:width="1dp"
+                android:color="#f0f0f0" />
+        </shape>
+    </item>
+</layer-list>
