+++ app/src/main/res/drawable/chatbox_bg.xml
@@ -1,24 +1,16 @@
-<?xml version="1.0" encoding="utf-8"?>
-<vector
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="360dp"
     android:height="60dp"
     android:viewportWidth="360"
-    android:viewportHeight="60"
-    >
-    <group>
-        <clip-path
-            android:pathData="M0 0H360V60H0V0Z"
-            />
-        <path
-            android:pathData="M0 0V60H360V0"
-            android:fillColor="#FFFFFF"
-            />
-        <path
-            android:pathData="M0 0H360V60H0V0Z"
-            android:strokeWidth="2"
-            android:strokeColor="#D8D8D8"
-            />
-    </group>
+    android:viewportHeight="60">
+  <path
+      android:pathData="M0,10C0,4.48 4.48,0 10,0H350C355.52,0 360,4.48 360,10V60H0V10Z"
+      android:fillColor="#ffffff"/>
+  <group>
+    <clip-path
+        android:pathData="M0,10C0,4.48 4.48,0 10,0H350C355.52,0 360,4.48 360,10V60H0V10Z"/>
+    <path
+        android:pathData="M-1,10C-1,3.92 3.92,-1 10,-1H350C356.08,-1 361,3.92 361,10H359C359,5.03 354.97,1 350,1H10C5.03,1 1,5.03 1,10H-1ZM360,60H0H360ZM-1,60V10C-1,3.92 3.92,-1 10,-1V1C5.03,1 1,5.03 1,10V60H-1ZM350,-1C356.08,-1 361,3.92 361,10V60H359V10C359,5.03 354.97,1 350,1V-1Z"
+        android:fillColor="#F0F0F0"/>
+  </group>
 </vector>
