+++ app/src/main/res/drawable/edit.xml
@@ -1,0 +1,11 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M1,6.75C1,3.574 3.574,1 6.75,1H16.75C19.926,1 22.5,3.574 22.5,6.75V16.75C22.5,19.926 19.926,22.5 16.75,22.5H6.75C3.574,22.5 1,19.926 1,16.75V6.75ZM13.413,6.83C13.843,6.401 14.539,6.4 14.968,6.83L16.678,8.539C17.104,8.965 17.108,9.654 16.688,10.086L15.547,11.256L12.268,7.977L13.413,6.83ZM11.685,8.561L14.971,11.847L10.875,16.049C10.565,16.368 10.139,16.547 9.695,16.547L7.825,16.547C7.356,16.547 6.981,16.156 7.001,15.687L7.081,13.788C7.098,13.375 7.27,12.983 7.562,12.691L11.685,8.561Z"
+      android:fillColor="#777777"
+      android:fillAlpha="0.466667"
+      android:fillType="evenOdd"/>
+</vector>
