+++ app/src/main/res/drawable/send.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="32dp"
+    android:height="32dp"
+    android:viewportWidth="32"
+    android:viewportHeight="32">
+  <path
+      android:pathData="M12.605,5.68L27.579,13.163C29.919,14.332 29.919,17.668 27.579,18.837L12.605,26.319C9.942,27.65 7.076,24.914 8.285,22.194L10.467,17.288C10.509,17.194 10.547,17.097 10.579,17L16,17C16.553,17 17,16.552 17,16C17,15.448 16.553,15 16,15L10.579,15C10.547,14.903 10.509,14.806 10.467,14.712L8.285,9.806C7.076,7.086 9.942,4.35 12.605,5.68ZM1.667,13.333C1.667,12.781 2.115,12.333 2.667,12.333H5.334C5.886,12.333 6.334,12.781 6.334,13.333C6.334,13.885 5.886,14.333 5.334,14.333H2.667C2.115,14.333 1.667,13.885 1.667,13.333ZM2.667,17.667C2.115,17.667 1.667,18.114 1.667,18.667C1.667,19.219 2.115,19.667 2.667,19.667H5.334C5.886,19.667 6.334,19.219 6.334,18.667C6.334,18.114 5.886,17.667 5.334,17.667H2.667Z"
+      android:fillColor="#777777"
+      android:fillType="evenOdd"/>
+</vector>
