+++ app/src/main/res/layout/fragment_chat.xml
@@ -4,8 +4,17 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/white"
-    tools:context=".ui.classui.chats.chat.ChatFragment">
+    android:background="#EEEEEE"
+    tools:context=".ui.classui.chats.ChatFragment">
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/chatRecyclerView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:scrollbars="vertical"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/toolbar" />
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
@@ -39,27 +48,15 @@
             android:layout_width="50dp"
             android:layout_height="50dp"
             android:background="@color/white"
-            android:src="@drawable/baseline_calendar_month_24"
+            android:src="@drawable/attendance_btn"
             android:contentDescription="To attendance page button"
             android:layout_gravity="end" />
     </androidx.appcompat.widget.Toolbar>
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/chatRecyclerView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:scrollbars="vertical"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/toolbar" />
-
-    <RelativeLayout
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/chatui"
-        android:layout_width="360dp"
+        android:layout_width="0dp"
         android:layout_height="60dp"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginTop="680dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -69,23 +66,31 @@
 
         <EditText
             android:id="@+id/chatText"
-            android:layout_width="wrap_content"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:ems="10"
+            android:layout_margin="10dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toStartOf="@id/chatSendButton"
+            android:background="@null"
             android:gravity="center_vertical"
             android:inputType="text"
-            android:hint="write your messages" />
+            android:hint="Write your messages"
+            android:fontFamily="@font/relay_noto_sans_medium"
+            android:textSize="16sp"
+            android:textColor="#777777" />
 
         <androidx.appcompat.widget.AppCompatImageButton
             android:id="@+id/chatSendButton"
-            android:layout_width="27dp"
-            android:layout_height="25dp"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentTop="true"
-            android:layout_marginLeft="312dp"
-            android:layout_marginTop="21dp"
-            android:background="#929292"
-            android:src="@android:drawable/ic_menu_send" />
-    </RelativeLayout>
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:layout_margin="10dp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:background="#FFFFFF"
+            android:src="@drawable/send" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
