+++ app/src/main/res/layout/fragment_chat_comment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.classui.chats.chat.ChatCommentFragment">
+    tools:context=".ui.classui.chats.ChatCommentFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
