+++ app/src/main/res/layout/item_message.xml
@@ -4,7 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:paddingVertical="16dp">
+    android:paddingVertical="16dp"
+    android:background="@drawable/chat_bg"
+    >
 
     <TextView
         android:id="@+id/messageText"
@@ -13,6 +15,8 @@
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
         android:textSize="20sp"
+        android:fontFamily="@font/relay_noto_sans_medium"
+        android:textColor="@color/black"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/nameAndTime"
@@ -20,11 +24,12 @@
 
     <androidx.appcompat.widget.AppCompatImageButton
         android:id="@+id/chatEditButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:src="@android:drawable/ic_menu_edit"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
+        android:layout_marginTop="10dp"
+        android:layout_marginEnd="10dp"
+        android:background="@color/white"
+        android:src="@drawable/edit"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -32,7 +37,10 @@
         android:id="@+id/nameAndTime"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:orientation="horizontal"
+        android:layout_marginStart="10dp"
+        android:text="TextView"
+        android:fontFamily="@font/relay_noto_sans_semibold"
+        android:textSize="16sp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent">
 
@@ -52,7 +60,6 @@
             android:textSize="10dp"
             android:layout_marginStart="8dp"
             android:text="TextView" />
-
     </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
